seperate API functions into seperate files in API folder
